fix(cli): Validate union base properties in examples
