+++ packages/cli/generation/ir-generator/src/examples/validateObjectExample.ts
@@ -9,6 +9,7 @@ import { getAllPropertiesForObject } from "../utils/getAllPropertiesForObject";
 import { ExampleViolation } from "./exampleViolation";
 import { getViolationsForMisshapenExample } from "./getViolationsForMisshapenExample";
 import { validateTypeReferenceExample } from "./validateTypeReferenceExample";
+import { validateObjectProperties } from "./validateObjectProperties";
 
 export function validateObjectExample({
     typeName,
@@ -32,79 +33,15 @@ export function validateObjectExample({
     workspace: FernWorkspace;
     breadcrumbs: string[];
 }): ExampleViolation[] {
-    if (!isPlainObject(example)) {
-        return getViolationsForMisshapenExample(example, "an object");
-    }
-
-    const violations: ExampleViolation[] = [];
-
-    const allPropertiesForObject = getAllPropertiesForObject({
+    return validateObjectProperties({
         typeName,
-        objectDeclaration: rawObject,
+        objectProperties: rawObject.properties,
+        extendedTypes: rawObject.extends,
+        file,
+        example,
         typeResolver,
-        definitionFile: file.definitionFile,
+        exampleResolver,
         workspace,
-        filepathOfDeclaration: file.relativeFilepath,
-        smartCasing: false
+        breadcrumbs
     });
-
-    const allPropertiesByWireKey = keyBy(allPropertiesForObject, (property) => property.wireKey);
-
-    // ensure required properties are present, we treat unknown as optional
-    const requiredProperties = allPropertiesForObject.filter(
-        (property) => !property.isOptional && property.resolvedPropertyType._type !== "unknown"
-    );
-    for (const requiredProperty of requiredProperties) {
-        // dont error on literal properties
-        if (
-            requiredProperty.resolvedPropertyType._type === "container" &&
-            requiredProperty.resolvedPropertyType.container._type === "literal"
-        ) {
-            continue;
-        }
-
-        if (example[requiredProperty.wireKey] == null) {
-            const propertyReference =
-                breadcrumbs.length > 0
-                    ? `${breadcrumbs.join(".")}.${requiredProperty.wireKey}`
-                    : requiredProperty.wireKey;
-            const message = `Example is missing required property "${propertyReference}"`;
-            violations.push({
-                message
-            });
-        }
-    }
-
-    // check properties on example
-    for (const [exampleKey, exampleValue] of Object.entries(example)) {
-        const propertyWithPath = allPropertiesByWireKey[exampleKey];
-        if (propertyWithPath == null) {
-            violations.push({
-                message: `Unexpected property "${exampleKey}"`
-            });
-        } else {
-            const definitionFile = getDefinitionFile(workspace, propertyWithPath.filepathOfDeclaration);
-            if (definitionFile == null) {
-                throw new Error("Service file does not exist for property: " + propertyWithPath.wireKey);
-            }
-            violations.push(
-                ...validateTypeReferenceExample({
-                    rawTypeReference: propertyWithPath.propertyType,
-                    example: exampleValue,
-                    file: constructFernFileContext({
-                        relativeFilepath: propertyWithPath.filepathOfDeclaration,
-                        definitionFile,
-                        casingsGenerator: file.casingsGenerator,
-                        rootApiFile: workspace.definition.rootApiFile.contents
-                    }),
-                    workspace,
-                    typeResolver,
-                    exampleResolver,
-                    breadcrumbs: [...breadcrumbs, `${exampleKey}`]
-                })
-            );
-        }
-    }
-
-    return violations;
 }
