+++ packages/cli/generation/ir-generator/src/utils/casingsGenerator.ts
@@ -1,0 +1,9 @@
+import { constructCasingsGenerator } from "../casings/CasingsGenerator";
+
+// Note: using this exported variable is NOT recommended, but its included for convenience
+// when the call-site doesn't care about the language nor special casing convention.
+export const CASINGS_GENERATOR = constructCasingsGenerator({
+    generationLanguage: undefined,
+    keywords: undefined,
+    smartCasing: false
+});
